Add `url` plugin option to support overriding the desired download URL
This is useful when you want to use this plugin with some kind of
custom build of Julia.  There's a little bit of annoying parsing we have
to do to fill out e.g. `bucket_host` and `path`, but overall I think
it's bearable.

I'm open to alternative implementations if we want to clean this up
further.